extend pipeline fixture inputs
