+++ examples/docker/multi_input/pipeline.dockerfile
