+++ examples/docker/multi_input/pipeline.yaml
